many changes across the whole tool
Common: tweaks and fixes
DBMP Editor:
improved VECTORVINYL support + optimization
fixed backups naming
Damage: will be continued, eventually. the file format seems much
trickier to understand than geometries though.
Collisions: removed ability to save edited collisions due to it causing
crashes. will be continued, eventually, if i can understand how to
calculate the FixUps.
Geometry Editor:
added a GUI with most of the functions from the CLI. a few functions,
such as map dump and replacement, still require using the CLI. some
functions may slightly vary between the CLI and GUI.
tweaks to the world map support, issues will still occur on certain
chunks
improved error handling
fixed hashed markers not being found in OBJ import
fixed tangents calculation crashing above 32k vertices (not that UC
models should ever get this high poly)
added crude TPK support for textures accompanying both car models and
world models (with automatic Z/Y/X/W/U LODs for the latter)
tweaks to the X360 platform support
improvements of platform support
added in-tool platform switching
beta support for legacy shaders from previous games
added a plugin for some compression formats (will require
added beta NFS ProStreet support (PC and X360 cars, platform switching
available, no compiling)
added partial NFS Carbon support (PC cars, broken, read-only)
and much more

+++ src/fr/ni240sx/ucgt/geometryFile/PartVisualizer.java
@@ -46,9 +46,9 @@ public class PartVisualizer extends Stage{
 				matMesh.setVertexFormat(VertexFormat.POINT_NORMAL_TEXCOORD);
 				
 				for (var v : m.verticesBlock.vertices) {
-					matMesh.getPoints().addAll((float)v.posX, (float)v.posY, (float)v.posZ);
-					matMesh.getNormals().addAll((float)v.normX, (float)v.normY, (float)v.normZ);
-					matMesh.getTexCoords().addAll((float)v.tex0U, 1-(float)v.tex0V);
+					matMesh.getPoints().addAll(v.posX, v.posY, v.posZ);
+					matMesh.getNormals().addAll(v.normX, v.normY, v.normZ);
+					matMesh.getTexCoords().addAll(v.tex0U, 1-v.tex0V);
 				}
 				
 				for (var tr : m.triangles) {
